+++ src/RecoveryServices/RecoveryServices.Backup/Cmdlets/Item/GetAzureRMRecoveryServicesBackupProtectableItem.cs
@@ -60,7 +60,27 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
         /// </summary>
         [Parameter(Mandatory = false, Position = 2, ParameterSetName = FilterParamSet,
             HelpMessage = ParamHelpMsgs.ProtectableItem.ItemType, ValueFromPipelineByPropertyName = false)]
+        [Parameter(Mandatory = false, Position = 2, ParameterSetName = IdParamSet,
+            HelpMessage = ParamHelpMsgs.ProtectableItem.ItemType, ValueFromPipelineByPropertyName = false)]
         public ProtectableItemType ItemType { get; set; }
+
+        /// <summary>
+        /// When this option is specified, only those items which belong to this name will be returned.
+        /// </summary>
+        [Parameter(Mandatory = false, ParameterSetName = FilterParamSet,
+            HelpMessage = ParamHelpMsgs.ProtectableItem.Name, ValueFromPipelineByPropertyName = false)]
+        [Parameter(Mandatory = false, ParameterSetName = IdParamSet,
+            HelpMessage = ParamHelpMsgs.ProtectableItem.Name, ValueFromPipelineByPropertyName = false)]
+        public string Name { get; set; }
+
+        /// <summary>
+        /// When this option is specified, only those items which belong to this server name will be returned.
+        /// </summary>
+        [Parameter(Mandatory = false, ParameterSetName = FilterParamSet,
+            HelpMessage = ParamHelpMsgs.ProtectableItem.ServerName, ValueFromPipelineByPropertyName = false)]
+        [Parameter(Mandatory = false, ParameterSetName = IdParamSet,
+            HelpMessage = ParamHelpMsgs.ProtectableItem.ServerName, ValueFromPipelineByPropertyName = false)]
+        public string ServerName { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -80,10 +100,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
                     string containerName = "";
                     Dictionary<UriEnums, string> keyValueDict = HelperUtils.ParseUri(ParentID);
                     containerName = HelperUtils.GetContainerUri(keyValueDict, ParentID);
-                    if (containerName.Split(new string[] { ";" }, System.StringSplitOptions.None)[0].ToLower() == "vmappcontainer")
-                    {
-                        backupManagementType = ServiceClientModel.BackupManagementType.AzureWorkload;
-                    }
+                    backupManagementType = ServiceClientModel.BackupManagementType.AzureWorkload;
                     string protectableItem = HelperUtils.GetProtectableItemUri(keyValueDict, ParentID);
                     if (protectableItem.Split(new string[] { ";" }, System.StringSplitOptions.None)[0].ToLower() == "sqlinstance" ||
                     protectableItem.Split(new string[] { ";" }, System.StringSplitOptions.None)[0].ToLower() == "sqlavailabilitygroupcontainer")
@@ -98,7 +115,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
                 }
                 else
                 {
-                    if(Container != null)
+                    if (Container != null)
                     {
                         string containerName = "";
                         backupManagementType = Container.BackupManagementType.ToString();
@@ -121,7 +138,6 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
                     }
                 }
 
-
                 WriteDebug("going to query service to get list of protectable items");
                 List<WorkloadProtectableItemResource> protectableItems =
                     ServiceClientAdapter.ListProtectableItem(
@@ -133,11 +149,30 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
 
                 if (ParameterSetName == FilterParamSet)
                 {
-                    string protectableItemType = ItemType.ToString();
-                    itemModels = itemModels.Where(itemModel =>
+                    if (ItemType != 0)
                     {
-                        return ((AzureWorkloadProtectableItem)itemModel).ProtectableItemType == protectableItemType;
-                    }).ToList();
+                        string protectableItemType = ItemType.ToString();
+                        itemModels = itemModels.Where(itemModel =>
+                        {
+                            return ((AzureWorkloadProtectableItem)itemModel).ProtectableItemType == protectableItemType;
+                        }).ToList();
+                    }
+
+                    if (Name != null)
+                    {
+                        itemModels = itemModels.Where(itemModel =>
+                        {
+                            return ((AzureWorkloadProtectableItem)itemModel).Name == Name;
+                        }).ToList();
+                    }
+
+                    if (ServerName != null)
+                    {
+                        itemModels = itemModels.Where(itemModel =>
+                        {
+                            return ((AzureWorkloadProtectableItem)itemModel).ServerName == ServerName;
+                        }).ToList();
+                    }
                 }
                 WriteObject(itemModels, enumerateCollection: true);
             });